Add code to match identification with a set of templates
This includes two functions, one will match one image, with one
template, the other takes a directory with templates, and match it with
one image. Both functions returns true or false.

This commit also includes templates for three different types of the
Norwegian driving license, togheter with one sample of each license to
test against.

The tests could be expanded with multiple images of the same
identification.